feat(daemon-client): implement HTTP API helpers for sandbox daemon
- Added a new file `daemon-client.ts` containing pure functions for interacting with the sandbox daemon's HTTP API, including health checks, command execution, and repository bootstrapping.
- Refactored existing Docker runner methods to utilize the new daemon client functions, improving code modularity and maintainability.
- Enhanced error handling and timeout management for daemon interactions.
